fix(tests): float64 mesh for test_ibm_pipe_poiseuille_3d (39% -> 8%)
ced7746 fixed the FVM Poiseuille CFL/float32 instability for the 2D
channel test (test_ibm_poiseuille_2d_channel_within_3pct) by plumbing
dtype=jnp.float64 through the mesh and BCs. It never touched the 3D
pipe test test_ibm_pipe_poiseuille_3d — a separate test, added in
5145871 and gpu+slow so it never ran in CI, so the miss was invisible.

The 3D test had the identical bug: _build_pipe_mesh built the mesh
with the default dtype=jnp.float32, so PISO ran float32 and hit the
same CFL-masked-by-reduction-noise convection instability — a 39.2%
profile error (the documented ~43% symptom).

Fix: give _build_pipe_mesh a dtype parameter and build the 3D
Poiseuille mesh + BCs float64, exactly as ced7746 did for 2D.
Verified: rel error 39.20% -> 8.24% (passes the 10% bar; 8.24% sits
in the docstring's stated 5-10% IBM-band signature range). The other
_build_pipe_mesh caller (static sphere drag) keeps the float32
default and is unaffected.

+++ tests/verification/test_fvm_ibm.py
@@ -125,7 +125,7 @@ def test_ibm_poiseuille_2d_channel_within_3pct():
 # 3D pipe Poiseuille via IBM cylinder + static sphere drag
 # ---------------------------------------------------------------------------
 
-def _build_pipe_mesh(N_cross, N_axial, R, L, *, margin=1.2):
+def _build_pipe_mesh(N_cross, N_axial, R, L, *, margin=1.2, dtype=jnp.float32):
     """Cubic-margin Cartesian box around a pipe of radius R, length L,
     periodic in z (axial direction)."""
     Lx = Ly = 2 * margin * R
@@ -133,7 +133,7 @@ def _build_pipe_mesh(N_cross, N_axial, R, L, *, margin=1.2):
     return make_cartesian_mesh_3d(
         N_cross, N_cross, N_axial, Lx, Ly, Lz,
         origin=(-Lx / 2, -Ly / 2, 0.0),
-        periodic_z=True,
+        periodic_z=True, dtype=dtype,
     )
 
 
@@ -153,7 +153,10 @@ def test_ibm_pipe_poiseuille_3d():
     L = 1.0
     nu = 0.005
     N_cross, N_axial = 24, 12
-    mesh = _build_pipe_mesh(N_cross, N_axial, R, L)
+    # float64 mesh — same fix ced7746 applied to the 2D channel test.
+    # PISO's explicit convection is CFL-limited; in float32 the GPU
+    # reduction noise masks the instability as a ~39% profile error.
+    mesh = _build_pipe_mesh(N_cross, N_axial, R, L, dtype=jnp.float64)
     dx = mesh.cartesian_spacing[0]
 
     def pipe_wall_sdf(x):
@@ -167,8 +170,8 @@ def test_ibm_pipe_poiseuille_3d():
         p = mesh.patch(name)
         nbf = int(p.owner.size)
         bcs[name] = VelocityBC(
-            u_wall=jnp.zeros((nbf, 3)),
-            F_through=jnp.zeros((nbf,)),
+            u_wall=jnp.zeros((nbf, 3), dtype=jnp.float64),
+            F_through=jnp.zeros((nbf,), dtype=jnp.float64),
         )
 
     f_steady = 0.005
